+++ src/main/java/moze_intel/projecte/gameObjs/container/inventory/TransmutationInventory.java
@@ -312,7 +312,7 @@ public class TransmutationInventory extends CombinedInvWrapper
 		}
 	}
 
-	public void addEmc(double value)
+	public void addEmc(long value)
 	{
 		provider.setEmc(provider.getEmc() + value);
 		
@@ -327,7 +327,7 @@ public class TransmutationInventory extends CombinedInvWrapper
 		}
 	}
 	
-	public void removeEmc(double value) 
+	public void removeEmc(long value) 
 	{
 		provider.setEmc(provider.getEmc() - value);
